Add POST /todos route and refactor mongoose
