+++ server/models/user.js
@@ -1,0 +1,10 @@
+const mongoose = require('mongoose');
+var User = mongoose.model('User', {
+    email: {
+        type: String,
+        required: true,
+        trim: true,
+        minlength: 1
+    }
+});
+module.exports = { User };
